Begin refactoring Jade engine into static library, and LevelEditor into application
This is the beginning of refactoring the code so that it is more platform
implementation friendly... Basically I'm trying to refactor my Window code
and implementation code so that the engine is a static library that gets linked
to by an application (in this case my level editor). This way I can separate the
two when shipping games.