Refactor major subsystems to avoid coupling. Read on for more details.
This commit is a huge change to the underlying structure of the engine.
Previously, the Scene was a singleton that was getting accessed through
various points in the code. Also, the entt library was being used throughout
the code which led to unhelpful error messages everytime an assertion failed.
To combat these, I undertook two large refactors:

  1. Decouple entt from all engine code (except systems)
  2. Remove the singleton behavior of scene (and pass to necessary subsystems instead)

1. First, I decoupled entt from as many subsystems as possible. I would like to
have it completely decoupled from the engine code, but that will be accomplished
in the future. To do this, I created the Entity class, which handles all interactions
between entt and the engine (aside from system interactions).

2. Secondly, I removed the singleton behavior from the Scene class. Since a scene
is an integral part of the engine, I am simply using that to my advantage. Only a few
subsystems really needed access to the scene, so it is now embedded in those systems as
a member variable. This means that there is now a complex start up sequence to make sure
dependencies get resolved (which is good, since we can easily tell who depends on whom).